Work in progress (Using Facts.  Should use Functions)
This isn't actually the right way to do this.  A HUP to the agent
doesn't send facts immediately, it waits until the catalog run finishes
to reload.

This is a problem for Step 4 because the agent no longer has a
certificate that it can use to submit facts.

The new plan is to migrate this work to a function to keep all of the
state server side.  When the master compiles the step 2 and step 4
catalog for the agent then it will record the fact the agent is on that
step.